% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasisBayeStat2025fall-package.R
\docType{package}
\name{oasisBayeStat2025fall-package}
\alias{oasisBayeStat2025fall}
\alias{oasisBayeStat2025fall-package}
\title{oasisBayeStat2025fall: A Streamlined R Realization of OASIS}
\description{
A compact R edition of Online Active Set method to Infer Spikes (OASIS) by Friedrich J. et al. (2017) \doi{10.1371/journal.pcbi.1005423}.
}
\references{
Friedrich J, Zhou P, Paninski L (2017). Fast online deconvolution of calcium 
imaging data. PLoS Comput Biol 13(3): e1005423.
}
\keyword{internal}
